ml-service/requirements.txt
numpy<2
pandas
scikit-learn
scikit-surprise
joblib